requirements.txt
python-dotenv
pandas
numpy
matplotlib
scikit-learn
pymongo
pymongo[srv]
certifi
dill
pyaml


#-e .